# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/moderation/queue": {
      "get": {
        "description": "Each item carries the automated reasons that flagged it, so a moderator sees why.",
        "operationId": "ModerationController_getQueue",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModerationQueueItemDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listings awaiting review, oldest first",
        "tags": ["moderation"]
      }
    },
    "/api/moderation/listings/{id}/approve": {
      "post": {
        "operationId": "ModerationController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveListingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish a listing from the queue",
        "tags": ["moderation"]
      }
    },
    "/api/moderation/media/{id}/preview": {
      "get": {
        "operationId": "ModerationController_moderationPreview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a short-lived private preview for an image awaiting review",
        "tags": ["moderation"]
      }
    },
    "/api/moderation/safety/cases": {
      "get": {
        "operationId": "ModerationController_safetyCases",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restricted child-safety case queue without image content",
        "tags": ["moderation"]
      }
    },
    "/api/moderation/safety/cases/{id}": {
      "get": {
        "description": "Returns no image, storage key, hash, or URL. Viewing the detail is itself recorded in the restricted audit log.",
        "operationId": "ModerationController_safetyCaseDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafetyCaseDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Restricted safety-case metadata and prior access history",
        "tags": ["moderation"]
      }
    },
    "/api/moderation/safety/cases/{id}/evidence-preview": {
      "post": {
        "description": "Not available to ordinary moderators. The justification is recorded before storage signs the URL.",
        "operationId": "ModerationController_safetyEvidencePreview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SafetyEvidenceAccessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Issue an audited, short-lived preview for held evidence",
        "tags": ["moderation"]
      }
    },
    "/api/moderation/safety/cases/{id}/report": {
      "post": {
        "description": "Only an open case can be reported. The external acknowledgement and actor justification are retained as restricted metadata.",
        "operationId": "ModerationController_reportSafetyCase",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportSafetyCaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record a safety case as reported through the approved channel",
        "tags": ["moderation"]
      }
    },
    "/api/moderation/safety/cases/{id}/release": {
      "post": {
        "description": "Does not publish the image or listing. The quarantined media returns to REVIEW_REQUIRED.",
        "operationId": "ModerationController_releaseSafetyCase",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveSafetyCaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Release a false-positive hold back to ordinary human review",
        "tags": ["moderation"]
      }
    },
    "/api/moderation/safety/cases/{id}/close": {
      "post": {
        "description": "Closing ends active handling; it does not delete evidence or release held media.",
        "operationId": "ModerationController_closeSafetyCase",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveSafetyCaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Close a reported safety case while preserving its legal hold",
        "tags": ["moderation"]
      }
    },
    "/api/moderation/listings/{id}/reject": {
      "post": {
        "description": "The reason is shown to the poster, so it must be written for them, not for internal notes.",
        "operationId": "ModerationController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectListingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject a listing",
        "tags": ["moderation"]
      }
    },
    "/api/moderation/listings/{id}/remove": {
      "post": {
        "operationId": "ModerationController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectListingDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a listing that is already published",
        "tags": ["moderation"]
      }
    },
    "/api/moderation/users/{id}/suspend": {
      "post": {
        "description": "Takes effect immediately: every session is revoked, so the account cannot keep acting on a token it already holds. Listings are left alone, because hiding the content of an account is a separate decision with its own trail.",
        "operationId": "ModerationController_suspendUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account suspended"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Suspend an account",
        "tags": ["moderation"]
      }
    },
    "/api/moderation/users/{id}/reinstate": {
      "post": {
        "description": "The account signs in again; revoked sessions are not restored.",
        "operationId": "ModerationController_reinstateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuspendUserDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lift a suspension",
        "tags": ["moderation"]
      }
    },
    "/api/moderation/media/{id}/block": {
      "post": {
        "description": "Removing a listing does not stop the same photograph being uploaded again a minute later. Blocking records both an exact and a perceptual hash, so a re-crop or a re-save of the same picture is refused too.",
        "operationId": "ModerationController_blockImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockImageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image blocked"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Refuse this image from now on",
        "tags": ["moderation"]
      }
    },
...
  },
  "info": {
    "title": "LocZ API",
    "description": "Location-first local discovery platform. Successful responses are wrapped in `{ success, data }`; errors use `{ success: false, error: { code, message } }`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://locz.in",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ModerationQueueItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "ownerName": {
            "type": "string"
          },
          "ownerPublishedCount": {
            "type": "number",
            "description": "Listings this owner already has published"
          },
          "cityName": {
            "type": "string"
          },
          "categoryName": {
            "type": "string"
          },
          "price": {
            "type": "object"
          },
          "moderationScore": {
            "type": "object"
          },
          "systemReasons": {
            "description": "Why the automated pass flagged this listing",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "imageCount": {
            "type": "number"
          },
          "reportCount": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "type",
          "ownerId",
          "ownerName",
          "ownerPublishedCount",
          "cityName",
          "categoryName",
          "systemReasons",
          "imageCount",
          "reportCount",
          "createdAt"
        ]
      },
      "ApproveListingDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "example": "Checked images, genuine listing"
          }
        }
      },
      "SafetyCaseAccessLogDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "actorId": {
            "type": "string"
          },
          "action": {
            "type": "string",
            "enum": [
              "CASE_VIEWED",
              "EVIDENCE_PREVIEW",
              "CASE_REPORTED",
              "HOLD_RELEASED",
              "CASE_CLOSED"
            ]
          },
          "justification": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": ["id", "actorId", "action", "justification", "createdAt"]
      },
      "SafetyCaseDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "mediaId": {
            "type": "string"
          },
          "listingId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["OPEN", "REPORTED", "RELEASED", "CLOSED"]
          },
          "mediaStatus": {
            "type": "string",
            "enum": [
              "PENDING_UPLOAD",
              "UPLOADED",
              "QUARANTINED",
              "SCANNING",
              "REVIEW_REQUIRED",
              "PROCESSING",
              "READY",
              "REJECTED",
              "LEGAL_HOLD",
              "FAILED"
            ]
          },
          "provider": {
            "type": "string"
          },
          "providerReference": {
            "type": "object",
            "nullable": true
          },
          "reasonCode": {
            "type": "string"
          },
          "reportReference": {
            "type": "object",
            "nullable": true
          },
          "resolutionNote": {
            "type": "object",
            "nullable": true
          },
          "openedAt": {
            "format": "date-time",
            "type": "string"
          },
          "reportedAt": {
            "type": "object",
            "nullable": true
          },
          "releasedAt": {
            "type": "object",
            "nullable": true
          },
          "closedAt": {
            "type": "object",
            "nullable": true
          },
          "accessHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SafetyCaseAccessLogDto"
            }
          }
        },
        "required": [
          "id",
          "mediaId",
          "listingId",
          "status",
          "mediaStatus",
          "provider",
          "reasonCode",
          "openedAt",
          "accessHistory"
        ]
      },
      "SafetyEvidenceAccessDto": {
        "type": "object",
        "properties": {
          "justification": {
            "type": "string",
            "example": "Verify provider match before statutory report 2026-07-26/04",
            "description": "Written to the restricted access log before a preview URL is issued. Never include image details."
          }
        },
        "required": ["justification"]
      },
      "ReportSafetyCaseDto": {
        "type": "object",
        "properties": {
          "reportReference": {
            "type": "string",
            "example": "REPORT-2026-07-26-004",
            "description": "Opaque acknowledgement from the approved reporting channel. Do not include image details."
          },
          "justification": {
            "type": "string",
            "example": "Submitted by the named safety officer through the approved reporting channel",
            "description": "Recorded in the restricted case access log."
          }
        },
        "required": ["reportReference", "justification"]
      },
      "ResolveSafetyCaseDto": {
        "type": "object",
        "properties": {
          "justification": {
            "type": "string",
            "example": "Confirmed false positive after the approved specialist review",
            "description": "Recorded in the restricted case access log. Never include image details or provider hashes."
          }
        },
        "required": ["justification"]
      },
      "RejectListingDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Contact details in the description instead of the contact fields"
          }
        },
        "required": ["reason"]
      },
      "SuspendUserDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Repeated fake listings after two warnings",
            "description": "Recorded in the audit trail — the reason is what makes the action reviewable"
          },
          "durationDays": {
            "type": "number",
            "example": 7,
            "description": "Days to suspend for. Omit for indefinite — but most suspensions should not be: a fortnight is a correction, a permanent ban for a first offence is a decision to lose a user rather than teach one.",
            "minimum": 1,
            "maximum": 365
          }
        },
        "required": ["reason"]
      },
      "BlockImageDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Photograph of ivory carving, listed as an antique bangle",
            "description": "Shown to the uploader if they try again, so it has to say something useful"
          },
          "category": {
            "type": "string",
            "example": "WILDLIFE",
            "description": "Matches the banned-keyword categories, so image and text refusals report together"
          }
        },
        "required": ["reason"]
      },
...
    }
  }
}